finished stateCardFilter
